compile

+++ minecraft-ecs/src/component.rs
@@ -2,8 +2,8 @@ use std::any::Any;
 
 
 
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
-pub struct ComponentId(pub u32);
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+pub struct ComponentId(pub u64);
 
 pub trait Component: Any + 'static {
     fn get_component_id(&self) -> ComponentId;
